+++ seatunnel-connectors-v2/connector-iceberg/src/main/java/org/apache/seatunnel/connectors/seatunnel/iceberg/catalog/IcebergCatalog.java
@@ -254,7 +254,7 @@ public class IcebergCatalog implements Catalog {
                             icebergDataTypeConvertor.toSeaTunnelType(
                                     name, nestedField.type().typeId());
                     return PhysicalColumn.of(
-                            name, seaTunnelType, (Long) null, true, null, nestedField.doc());
+                            name, seaTunnelType, null, true, null, nestedField.doc());
                 });
 
         List<String> partitionKeys =
